+++ CHANGELOG.rst
@@ -6,8 +6,9 @@ dev
 
 **Improvements**
 
-- Added ``PeriodInstancesAPI.relative()`` for relative queries
+- Add ``PeriodInstancesAPI.relative()`` for relative queries
 - Add ``modified-at-latest`` parameter for ``eq.instances.relative()``
+- Implement absolute forecasts for instances ``eq.instances.absolute()``
 
 **Bugfixes**